Revert "Add versioned indexes and improve full name search in OpenSearch (#1334)"
This reverts commit 3d895dc5cd28d138dc54990a7a28aafb6ff4908e.
