+++ backend/src/services/aws.ts
@@ -1,5 +1,4 @@
 import AWS, { SQS } from 'aws-sdk'
-import { trimUtf8ToMaxByteLength } from '@crowd/common'
 import { COMPREHEND_CONFIG, IS_DEV_ENV, KUBE_MODE, S3_CONFIG, SQS_CONFIG } from '../conf'
 
 let sqsInstance
@@ -99,6 +98,22 @@ if (KUBE_MODE) {
     process.env.AWS_SECRET_ACCESS_KEY !== undefined
       ? new AWS.Comprehend()
       : undefined
+}
+
+const trimUtf8ToMaxByteLength = (utf8Str: string, maxByteLength: number): string => {
+  if (Buffer.byteLength(utf8Str, 'utf8') > maxByteLength) {
+    // this will get us close but some characters could be multibyte encoded so we might need to trim a bit more
+    utf8Str = utf8Str.slice(0, maxByteLength)
+  }
+
+  // trim till we get to the requested byte length or lower (if we cut multibyte character)
+  let byteLength = Buffer.byteLength(utf8Str, 'utf8')
+  while (byteLength > maxByteLength) {
+    utf8Str = utf8Str.slice(0, -1)
+    byteLength = Buffer.byteLength(utf8Str, 'utf8')
+  }
+
+  return utf8Str
 }
 
 const ALLOWED_MAX_BYTE_LENGTH = 5000
