+++ services/apps/search_sync_worker/src/types.ts
@@ -4,13 +4,6 @@ export enum OpenSearchIndex {
   ORGANIZATIONS = 'organizations',
 }
 
-// Keeps track of version numbers for all OpenSearch indexes, aiding in managing documents.
-// for eg: members_v1, activities_v1, etc.
-export const IndexVersions = new Map<OpenSearchIndex, number>()
-IndexVersions.set(OpenSearchIndex.MEMBERS, 1)
-IndexVersions.set(OpenSearchIndex.ACTIVITIES, 1)
-IndexVersions.set(OpenSearchIndex.ORGANIZATIONS, 1)
-
 const prefixedMapping = {
   dynamic_templates: [
     // https://opensearch.org/docs/latest/field-types/supported-field-types/string/
@@ -21,7 +14,6 @@ const prefixedMapping = {
         path_match: '.*',
         mapping: {
           type: 'text',
-          analyzer: 'lowercase_keyword_analyzer',
         },
       },
     },
@@ -574,24 +566,6 @@ const prefixedMapping = {
       },
     },
   ],
-}
-
-const prefixedSettings = {
-  analysis: {
-    analyzer: {
-      lowercase_keyword_analyzer: {
-        type: 'custom',
-        tokenizer: 'keyword',
-        filter: ['lowercase'],
-      },
-    },
-  },
-}
-
-export const OPENSEARCH_INDEX_SETTINGS: Record<OpenSearchIndex, unknown> = {
-  [OpenSearchIndex.MEMBERS]: prefixedSettings,
-  [OpenSearchIndex.ACTIVITIES]: prefixedSettings,
-  [OpenSearchIndex.ORGANIZATIONS]: prefixedSettings,
 }
 
 export const OPENSEARCH_INDEX_MAPPINGS: Record<OpenSearchIndex, unknown> = {
