+++ services/libs/common/src/index.ts
@@ -9,4 +9,3 @@ export * from './strings'
 export * from './types'
 export * from './requestThrottler'
 export * from './rawQueryParser'
-export * from './byteLength'
